add an integration test for OTLP HTTP span exporter (#4)
Add an integration test for the OpenTelemetry OTLP HTTP span exporter.

The plugin is also simplified so only a single exporter is configured because otherwise it would greatly complicate the new configuration options. (I may revisit this in the future if and when there are non-OpenTelemetry exporters in play.)

OpenTelemetry usually pulls configuration from environment variables. This does not really match up well with the Pants configuration system, especially since it would involve querying the engine for those environment variables using `os.environ` which may not reflect the user's environment if pantsd is involved.

Instead, I chose to use the Pants configuration system and incorporated the help text from the OpenTelemetry project.