apply workunit metadata as attributes to otel spans (#13)
Apply workunit level and metadata as attributes on generated OpenTelemetry spans. The attributes are prefixed with `pantsbuild.workunit` as a namespace. The metadata dictionary is flattened into OpenTelemetry attributes with keys emitted as `pantsbuild.workunit.metdata.KEY` attribitues. Only primitive values are handled currently.

Also a bug in construction of the `FrozenDict` used for workunit metadata since complex values were not "deeply" frozen.