distribute the plugin as a PEX and not as a wheel on (private) PyPI (#28)
Distribute the plugin in PEX form and not as a wheel. While testing the end user setup workflow for the plugin, Pants was taking 30-45 seconds to resolve the plugin from a private PyPI repository. This is not an acceptable end user experience.

Instead, the end user will now expand the plugin's pex file into a venv, and then update `[GLOBAL].pythonpath` to include that venv's `site-packages` directory. In testing of this solution, there is a negligible time cost for that in comparison to the abysmal plugin resolution time found when using wheels and a private PyPI repository.

The integration test is updated to make use of the PEX format.