remove old stable pants versions to simplify testing (#35)
Remove support for old stable Pants versions 2.23 and 2.24. This will speed up testing by not testing Pants versions which are not supported any more upstream.

Note: Support for Pants v2.21 is kept since an initial user of this plugin is still on Pants v2.21. Once support for 2.21 is removed, the plugin can be further simplified to only support Python 3.11 and not both 3.9 and 3.11. (Pants v2.25 switched to using Python 3.11.)