produce a single wheel artifact for distribution (#39)
The plugin works with multiple Pants versions as is (and should adapt
itself as necessary). So just distribute a single wheel artifact based
on the resolve for the latest Pants version which is supported. The
Pants dependency is excluded from the wheel so there should be no
conflicts with users varying the Pants version.

Note: This probably only works for Pants v2.25 and higher since that is
when the switch to Python 3.11 occurred.