prevent in-flight spans from being dropped during session shutdown (#53)
As described in
https://github.com/shoalsoft/shoalsoft-pants-opentelemetry-plugin/issues/11,
the existing plugin code appears to be shutting down the OpenTelemetry
exporter while it is still trying to export spans.

This PR adds a few fixes:
- The `Processor.finish` method has a new `timeout` parameter which
allows the caller to inform the underlying libraries how long they have
to finish emitting spans.
- `OpenTelemetryProcessor.finish` now explicitly flushes the underlying
exporter and applies the configured timeout before shutting down the
exporter.
- To support async completion of workunit processing, workunits should
not be processed on the thread which receives them from the engine. The
new `SingleThreadedProcessor` class implements `Processor` and wraps
another `Processor` and drives it on a single, separate thread. (Also,
the `opentelemetry-python` support should be thread-safe, but in case it
isn't, `SingleThreadedProcessor` will also handle that case.)