pass repository explicitly to `gh` commad in publish workflow (#71)
Apparently `gh` tries in infer the repository to use by running `git` if
the `--repo` option is not provided. However, the error message is an
obtuse `failed to run git: fatal: not a git repository (or any of the
parent directories): .git` without any explanation for why `git` was
being invoked.

Pass the repository explicitly into `gh` so that we can still avoid
checking out the repo in the validation job.