add `--shoalsoft-opentelemetry-async-completion` option (#82)
Add a new `--shoalsoft-opentelemetry-async-completion` option to control
async completion of work unit processing. By default, the option is
enabled and Pants will not wait for the work unit handler to signal
completion. This may result in loss of work units though if the handler
is not able to transmit them in time. The option gives the user the
choice.