add subprocess processor to offload gRPC-related work (#85)
As described in
https://github.com/shoalsoft/shoalsoft-pants-opentelemetry-plugin/issues/84,
the gRPC C library has documented fork safety issues due to being
multi-threaded. It does not operate properly in processes which fork
(especially not in a project like Pantsbuild which spawns off build
actions regularly). Prior testing was in the context of (contrived)
integration tests and did not trigger the fork safety issue. Trying to
use this plugin in production during CI did trigger the fork safety
issues though.

Solve the issue by offloading work unit processing to a subprocess when
the gRPC exporter is configured. The new `SubprocessProcessor` offloads
work unit processing to a subprocess. That subprocess is where the gRPC
C library is initialized and the subprocess never forks which avoids the
fork safety issue.

A prior attempt to solve the problem used `multiprocessing` but that
attempt had its own issues with fork safety. In the end, the simpler
solution was to just spawn a dedicated Python interpreter and
communicate using pickled objects over stdio pipes.