move processor initialization into callback (#96)
The existing logic initialized the work unit processor outside of the
work unit callback and passed it into the callback as a parameter. This
caused the processor to be cached and only initialized once given how
the engine works. The correct logic is to initialize the processor
inside the work unit callback which is what is called each session.

This fixes a bug with tracing spans not being sent after the first
session to run when running with pantsd.