add support to output links to traces in a tracing system (#106)
As described in
https://github.com/shoalsoft/shoalsoft-pants-opentelemetry-plugin/issues/103,
it is useful to output a link to the trace in the
OpenTelemetry-compatible tracing system collecting the work unit spans.

This PR adds a new `[shoalsoft-opentelemetry].trace_link_template`
option which, when set, defines a template defining how to output trace
links to the log at INFO level.