updates Pants versions used for integration testing (#122)
Changes to the Pants versions used for integration testing:

- 2.30.x: Add Pants 2.30.0rc0 to the integration test. This version is
also used for the plugin's own Pants version.
- 2.29.x: Upgrade to Pants v2.29.1.
- 2.28.x: Upgrade to Pants v2.28.1.
- 2.27.x / 2.26.x: Remove testing of the plugin with Pants v2.27.x and
v2.26.x since those versions are old now.

Also upgrades Pex to v2.69.0 and the Pip version to be `latest` instead
of pinning it to `25.0` which caused issues with v25.0 not being able to
recognize `CPython<3.14` specifiers.