avoid hard-coding the full Pants versions in the integration test (#123)
Avoid hard-coding the full versions of the Pants versions used for
integration testing. Instead, just query the Pants PEX in use for
particular major/minor version for the full version. This needs to be
done because the full version needs to be specified in the `pants.toml`
used for the test run.